LayoutTests/http/tests/websocket/tests/hixie76/handshake-fail-by-cross-origin-expected.txt

CONSOLE MESSAGE: Error during WebSocket handshake: origin mismatch: http://127.0.0.1:8000 != http://example.com
Make sure Web Socket connection failed if origin mismatches.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

document.domain=127.0.0.1
ws.url=ws://localhost:8880/websocket/tests/hixie76/fixed-origin
PASS connected is false
PASS origin is undefined.
PASS successfullyParsed is true

TEST COMPLETE
